Add files via upload
This Python code implements a Support Vector Regression (SVR) model for a regression task. SVR is a supervised learning algorithm that excels at finding a hyperplane that best separates the target variable (Y) from the features (X) in a high-dimensional feature space. In the context of regression, the goal is to minimize the prediction error between the actual values of Y and the values predicted by the model.